responsive hero section
