+++ src/pages/security/_hero.js
@@ -1,47 +1,16 @@
 import React from 'react'
 import { graphql, useStaticQuery } from 'gatsby'
-import styled from 'styled-components'
-import { BackgroundImage , Header } from 'components/elements'
-import { Container } from 'components/containers'
-import device from 'themes/device'
-
-const StyledContainer = styled(Container)`
-    flex-direction: column;
-    height: 100%;
-    width: 60%;
-
-    @media ${device.tablet} {
-        padding: 0 2rem;
-        align-items: flex-start;
-        width: 100%;
-    }
-`
-
-const StyledHeader = styled(Header)`
-    margin-bottom: 1.6rem;
-    line-height: 8rem;
-    font-size: 6.4rem;
-
-    @media ${device.tablet} {
-        font-size: 40px;
-        line-height: 50px;
-    }
-`
-
-const Subheadline = styled(Header)`
-    font-size: 2.4rem;
-    font-weight: normal;
-    line-height: 36px;
-
-    @media ${device.tablet} {
-        font-size: 18px;
-        line-height: 26px;
-    }
-`
+import HeroComponent from './components/_hero_component'
+import { Show } from 'components/containers'
 
 const query = graphql`
     query {
-        image: file(relativePath: { eq: "security/security-overlay.jpg" }) {
+        hero_background_desktop: file(
+            relativePath: { eq: "security/security-overlay-desktop.jpg" }
+        ) {
+            ...heroImage
+        }
+        hero_background_mobile: file(relativePath: { eq: "security/security-overlay-mobile.jpg" }) {
             ...heroImage
         }
     }
@@ -51,26 +20,22 @@ const Hero = () => {
     const data = useStaticQuery(query)
 
     return (
-        <BackgroundImage
-            data={data.image}
-            alt="deriv security"
-            style={{
-                height: '90vh',
-                width: '100vw',
-                backgroundSize: `cover`,
-                maxWidth: '100%',
-            }}
-        >
-            <StyledContainer>
-                <StyledHeader as="h1" color="white" align="center">
-                    Bug bounty program
-                </StyledHeader>
+        <>
+            <Show.Desktop>
+                <HeroComponent
+                    background_data={data['hero_background_desktop']}
+                    background_height="70vh"
+                />
+            </Show.Desktop>
 
-                <Subheadline as="h3" color="white" align="center">
-                    Security is a collaboration. Report bugs and be rewarded.
-                </Subheadline>
-            </StyledContainer>
-        </BackgroundImage>
+            <Show.Mobile>
+                <HeroComponent
+                    background_data={data['hero_background_mobile']}
+                    background_height="60vh"
+                    background_dark="0.3"
+                />
+            </Show.Mobile>
+        </>
     )
 }
