+++ src/pages/security/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import Rewards from './rewards'
+import Hero from './_hero'
 import { localize, WithIntl } from 'components/localization'
 import Layout from 'components/layout/layout'
 import { SEO } from 'components/containers'
@@ -10,6 +11,7 @@ const Security = () => {
             <SEO title={localize('Some title')} description={localize('Some description')} />
             Hello World!
             <Rewards />
+            <Hero />
         </Layout>
     )
 }
